refactor(cart): Remove deprecated cart query files to streamline functionality
